Adicionando informações nas classes
