+++ solutions/devsprint-michel-ferreira-1/src/main/kotlin/com/devpass/challengehexagonal/domain/model/Account.kt
@@ -1,10 +1,7 @@
 package com.devpass.challengehexagonal.domain.model
 
-import java.time.LocalDate
-import java.util.UUID
-
-class Client(
-    id: UUID,
-    name: String,
-    birthDate: LocalDate?
+class Account(
+    number: Int,
+    balance_available: Double,
+    transaction_history: List<Transaction>
 )
