+++ solutions/devsprint-michel-ferreira-1/src/main/kotlin/com/devpass/challengehexagonal/domain/model/Client.kt
@@ -1,4 +1,10 @@
 package com.devpass.challengehexagonal.domain.model
 
-class Client {
-}
+import java.time.LocalDate
+import java.util.UUID
+
+class Client(
+    id: UUID = UUID.randomUUID(),
+    name: String,
+    birthDate: LocalDate?
+)
