+++ solutions/devsprint-michel-ferreira-1/src/main/kotlin/com/devpass/challengehexagonal/domain/model/Transaction.kt
@@ -1,10 +1,9 @@
 package com.devpass.challengehexagonal.domain.model
 
-import java.time.LocalDate
-import java.util.UUID
+import java.time.LocalDateTime
 
-class Client(
-    id: UUID = UUID.randomUUID(),
-    name: String,
-    birthDate: LocalDate?
+class Transaction(
+    establishment: String,
+    createdAt: LocalDateTime = LocalDateTime.now(),
+    value: Double
 )
