+++ solutions/devsprint-michel-ferreira-1/src/main/kotlin/com/devpass/challengehexagonal/domain/model/Violation.kt
@@ -1,11 +1,6 @@
 package com.devpass.challengehexagonal.domain.model
 
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.util.UUID
+enum class Violation{
+    INSUFFICIENT_FOUNDS
+}
 
-class Transaction(
-    establishment: String,
-    createdAt: LocalDateTime = LocalDateTime.now(),
-    value: Double
-)
